fix check_not_found_exception
